# teacher/backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TIDE Teacher API Contract
  version: 0.10.0
  description: >-
    教师端公共合同。包含教师端任务、文件和 My TIDE 安全读取；
    各模块实现状态见 contracts/README。任务、G01 状态和外部消息直接使用共享 PostgreSQL 契约表；
    教师积分与逐课积分直接读取世文当前聚合视图，不在教师端重算。
servers:
  - url: http://localhost:3000
    description: local backend
...
paths:
...
  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: 返回世文已为当前老师创建的共享任务
      operationId: listTeacherTasks
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 共享 assignment 任务列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskListResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/tasks/{taskInstanceId}:
    get:
      tags: [Tasks]
      summary: 获取统一任务上下文
      operationId: getTaskContext
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
      responses:
        '200':
          description: 共享任务与本地执行步骤组成的上下文
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskContext' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tasks/{taskInstanceId}/view:
    post:
      tags: [Tasks]
      summary: 老师主动打开任务详情后标记已查看
      description: 任务列表曝光不得调用此接口。
      operationId: viewTask
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MutationMeta' }
      responses:
        '200': { $ref: '#/components/responses/TaskMutation' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/tasks/{taskInstanceId}/start:
    post:
      tags: [Tasks]
      summary: 开始任务
      operationId: startTask
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MutationMeta' }
      responses:
        '200': { $ref: '#/components/responses/TaskMutation' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/tasks/{taskInstanceId}/progress:
    put:
      tags: [Tasks]
      summary: 保存任务步骤进度
      operationId: saveTaskProgress
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveProgressRequest' }
      responses:
        '200': { $ref: '#/components/responses/TaskMutation' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/tasks/{taskInstanceId}/submissions:
    post:
      tags: [Tasks]
      summary: 提交任务尝试
      operationId: submitTask
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitTaskRequest' }
      responses:
        '202': { $ref: '#/components/responses/TaskMutation' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/tasks/{taskInstanceId}/validation:
    get:
      tags: [Tasks]
      summary: 查询最近一次验证结果
      operationId: getTaskValidation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
      responses:
        '200':
          description: 验证结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationResult' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tasks/{taskInstanceId}/teacher-photo:
    post:
      tags: [TeacherPhoto]
      summary: 提交首课摄像头画面，完成四项可见内容检测和私有保存
      operationId: submitTeacherPhoto
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [photo]
              properties:
                photo: { type: string, format: binary }
      responses:
        '201':
          description: 检测与保存状态；四项通过后不会再要求教师重拍
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherPhotoRun' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [TeacherPhoto]
      summary: 查询当前首课摄像头画面检测和保存状态
      operationId: getTeacherPhoto
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
      responses:
        '200':
          description: 最新摄像头画面检测状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherPhotoRun' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tasks/{taskInstanceId}/teacher-photo/content:
    get:
      tags: [TeacherPhoto]
      summary: 读取已鉴权的画面检测证据
      operationId: downloadTeacherPhoto
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
      responses:
        '200':
          description: 后台处理并私有保存的画面检测证据
          headers:
            Cache-Control:
              schema: { type: string, const: 'private, no-store' }
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tasks/{taskInstanceId}/retry:
    post:
      tags: [Tasks]
      summary: 创建下一次任务尝试
      operationId: retryTask
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskInstanceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RetryTaskRequest' }
      responses:
        '200': { $ref: '#/components/responses/TaskMutation' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/support-tickets:
    get:
      tags: [SupportTickets]
      summary: 读取当前教师工单并在本次刷新中同步运营回复状态
      operationId: listSupportTickets
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 工单列表与未读数
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicketList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [SupportTickets]
      summary: 创建教师工单
      operationId: createSupportTicket
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/CreateSupportTicket' }
      responses:
        '201':
          description: 工单已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicket' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: 单张截图超过 8 MB }
        '415': { description: 截图格式不支持 }
  /api/v1/support-tickets/{ticketId}:
    get:
      tags: [SupportTickets]
      summary: 读取当前教师的一条工单
      operationId: getSupportTicket
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 工单详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicket' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/support-tickets/{ticketId}/read:
    post:
      tags: [SupportTickets]
      summary: 将当前最新运营回复标为已读
      operationId: markSupportTicketRead
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 已更新工单
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicket' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/support-tickets/{ticketId}/messages:
    post:
      tags: [SupportTickets]
      summary: 教师在未关闭工单中继续补充
      operationId: replySupportTicket
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ReplySupportTicket' }
      responses:
        '200':
          description: 消息已追加
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicket' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/support-tickets/{ticketId}/resolve:
    post:
      tags: [SupportTickets]
      summary: 教师确认已解决并由 TIDE 关闭工单
      operationId: resolveSupportTicket
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [rowVersion]
              properties:
                rowVersion: { type: integer, minimum: 1 }
      responses:
        '200':
          description: 工单已关闭并触发图片清理
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SupportTicket' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/support-tickets/{ticketId}/images/{fileId}:
    get:
      tags: [SupportTickets]
      summary: 鉴权读取当前教师工单图片
      operationId: getSupportTicketImage
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 私有图片二进制
          content:
            image/jpeg: { schema: { type: string, contentEncoding: binary } }
            image/png: { schema: { type: string, contentEncoding: binary } }
            image/webp: { schema: { type: string, contentEncoding: binary } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TaskInstanceId:
      name: taskInstanceId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 255 }
    TicketId:
      name: ticketId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    TaskMutation:
      description: 状态变更已持久化
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TaskMutationResponse' }
    Unauthorized:
      description: 未认证或验签失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    BadRequest:
      description: 请求字段格式错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: 资源不存在或不属于当前教师
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: 幂等负载冲突、状态版本冲突或非法状态迁移
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ValidationError:
      description: 请求字段或任务输出不符合合同
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServiceUnavailable:
      description: 数据源或外部依赖暂时不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    MutationMeta:
      type: object
      additionalProperties: false
      required: [commandId, expectedStateVersion]
      properties:
        commandId: { type: string, minLength: 8, maxLength: 128 }
        expectedStateVersion: { type: integer, minimum: 1 }
    TaskListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TaskSummary' }
    TaskSummary:
      type: object
      additionalProperties: false
      required: [taskInstanceId, taskCode, kind, status, stateVersion, display, title, priority, availableAt, dueAt, completedAt, dataOrigin]
      properties:
        taskInstanceId: { type: string }
        taskCode: { type: string }
        kind: { type: string, enum: [FIXED_GROWTH, PERSONALIZED_IMPROVEMENT] }
        status: { $ref: '#/components/schemas/TaskStatus' }
        stateVersion: { type: integer, minimum: 1 }
        display: { $ref: '#/components/schemas/TaskDisplayMetadata' }
        title: { type: string }
        priority: { type: string, enum: [P0, P1, P2, P3] }
        availableAt: { type: [string, 'null'], format: date-time }
        dueAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        dataOrigin: { type: string, enum: [REAL, DERIVED_REAL, MOCK, MOCK_SIMULATION, MOCK_PROXY] }
    TaskDisplayMetadata:
      type: object
      additionalProperties: false
      required: [stageKey, sequence, points, estimatedMinutes]
      properties:
        stageKey: { type: [string, 'null'] }
        sequence: { type: [integer, 'null'], minimum: 1 }
        points: { type: number, minimum: 0 }
        estimatedMinutes: { type: [integer, 'null'], minimum: 1 }
    TaskContext:
      type: object
      additionalProperties: false
      required: [taskInstanceId, taskCode, kind, templateVersion, executionContractVersion, status, stateVersion, display, content, steps, progress, capabilities, assignment, availableAt, dueAt, completedAt, dataOrigin]
      properties:
        taskInstanceId: { type: string }
        taskCode: { type: string }
        kind: { type: string, enum: [FIXED_GROWTH, PERSONALIZED_IMPROVEMENT] }
        templateVersion: { type: integer, minimum: 1 }
        executionContractVersion: { type: string }
        status: { $ref: '#/components/schemas/TaskStatus' }
        stateVersion: { type: integer, minimum: 1 }
        display: { $ref: '#/components/schemas/TaskDisplayMetadata' }
        content: { $ref: '#/components/schemas/TaskContent' }
        steps:
          type: array
          items: { $ref: '#/components/schemas/TaskStep' }
        progress: { $ref: '#/components/schemas/TaskProgress' }
        capabilities:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/StepType' }
        assignment: { $ref: '#/components/schemas/SharedAssignmentContext' }
        availableAt: { type: [string, 'null'], format: date-time }
        dueAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        dataOrigin: { type: string, enum: [REAL, DERIVED_REAL, MOCK, MOCK_SIMULATION, MOCK_PROXY] }
    TaskContent:
      type: object
      additionalProperties: false
      required: [title, why, whatToDo, completionStandard, outcome, language]
      properties:
        title: { type: string }
        why: { type: string }
        whatToDo: { type: string }
        completionStandard: { type: string }
        outcome: { type: string }
        language: { type: string }
    SharedAssignmentContext:
      type: object
      additionalProperties: false
      required: [assignmentId, teacherSafeReason, teacherSafeFacts, relatedCourses, reminderNotificationId, priority, dueAt]
      properties:
        assignmentId: { type: string }
        teacherSafeReason: { type: string }
        teacherSafeFacts:
          type: array
          items: { $ref: '#/components/schemas/TaskTeacherSafeFact' }
        relatedCourses:
          type: array
          items: { $ref: '#/components/schemas/TaskRelatedCourse' }
        reminderNotificationId: { type: [string, 'null'] }
        priority: { type: string, enum: [P0, P1, P2, P3] }
        dueAt: { type: [string, 'null'], format: date-time }
    TaskTeacherSafeFact:
      type: object
      additionalProperties: false
      required: [label, labelZh, value, valueZh]
      properties:
        label: { type: string }
        labelZh: { type: [string, 'null'] }
        value: { type: string }
        valueZh: { type: [string, 'null'] }
...
    TaskStep:
      type: object
      additionalProperties: false
      required: [stepKey, position, type, title, config]
      properties:
        stepKey: { type: string }
        position: { type: integer, minimum: 1 }
        type: { $ref: '#/components/schemas/StepType' }
        title: { type: string }
        config: { type: object, additionalProperties: true }
    TaskProgress:
      type: object
      required: [currentStepKey, percent, steps]
      properties:
        currentStepKey: { type: [string, 'null'] }
        percent: { type: integer, minimum: 0, maximum: 100 }
        steps:
          type: array
          items:
            type: object
            required: [stepKey, status, percent]
            properties:
              stepKey: { type: string }
              status: { type: string, enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, FAILED] }
              percent: { type: integer, minimum: 0, maximum: 100 }
    StepType:
      type: string
      enum: [VIDEO, DOCUMENT, QUIZ, CHECKLIST, UPLOAD, DEVICE_CHECK, EXTERNAL_TRAINING, CUSTOM]
    TaskStatus:
      type: string
      enum: [ASSIGNED, VIEWED, IN_PROGRESS, SUBMITTED, UNDER_REVIEW, COMPLETED, FAILED, EXPIRED, WAIVED, CANCELLED]
    SaveProgressRequest:
      type: object
      additionalProperties: false
      required: [commandId, expectedStateVersion, stepKey, percent, progress]
      properties:
        commandId: { type: string, minLength: 8, maxLength: 128 }
        expectedStateVersion: { type: integer, minimum: 1 }
        stepKey: { type: string }
        percent: { type: integer, minimum: 0, maximum: 100 }
        progress: { type: object, additionalProperties: true }
    SubmitTaskRequest:
      type: object
      additionalProperties: false
      required: [commandId, expectedStateVersion, attemptId, outputs]
      properties:
        commandId: { type: string, minLength: 8, maxLength: 128 }
        expectedStateVersion: { type: integer, minimum: 1 }
        attemptId: { type: string, format: uuid }
        outputs:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/StepOutput' }
    StepOutput:
      type: object
      additionalProperties: false
      required: [stepKey, outputType, value]
      properties:
        stepKey: { type: string }
        outputType: { type: string, enum: [QUIZ, CHECKLIST, FILE, DEVICE_CHECK, EXTERNAL_PROOF, CUSTOM] }
        value: { type: object, additionalProperties: true }
    RetryTaskRequest:
      type: object
      additionalProperties: false
      required: [commandId, expectedStateVersion]
      properties:
        commandId: { type: string, minLength: 8, maxLength: 128 }
        expectedStateVersion: { type: integer, minimum: 1 }
        reasonCode: { type: string, maxLength: 128 }
    TaskMutationResponse:
      type: object
      required: [accepted, taskInstanceId, status, stateVersion]
      properties:
        accepted: { type: boolean, const: true }
        taskInstanceId: { type: string }
        status: { $ref: '#/components/schemas/TaskStatus' }
        stateVersion: { type: integer, minimum: 1 }
        step:
          type: object
          additionalProperties: false
          required: [stepKey, status, percent]
          properties:
            stepKey: { type: string }
            status: { type: string, enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, FAILED] }
            percent: { type: integer, minimum: 0, maximum: 100 }
            result: { type: object, additionalProperties: true }
        validation: { $ref: '#/components/schemas/ValidationResult' }
    ValidationResult:
      type: object
      required: [status, resultCode, teacherMessage]
      properties:
        status: { type: string, enum: [PENDING, UNDER_REVIEW, PASSED, FAILED, ERROR] }
        resultCode: { type: [string, 'null'] }
        teacherMessage: { type: [string, 'null'] }
...
    TeacherPhotoCheck:
      type: object
      additionalProperties: false
      required: [id, title, status, message, suggestion]
      properties:
        id: { type: string, enum: [camera_angle, lighting, background, dressing] }
        title: { type: string }
        status: { type: string, enum: [pass, fail, uncertain] }
        message: { type: string }
        suggestion: { type: string }
    TeacherPhotoRun:
      type: object
      additionalProperties: false
      required: [photoRunId, status, decision, criteriaVersion, teacherMessage, submittedAt, checkedAt, checks, finalPhoto]
      properties:
        photoRunId: { type: string, format: uuid }
        status:
          type: string
          enum: [UPLOADING, CHECKING, RETRY_REQUIRED, BEAUTIFYING, READY, UNDER_REVIEW, PROCESSING_FAILED]
        decision:
          oneOf:
            - { type: string, enum: [PASS, RETRY, ERROR] }
            - { type: 'null' }
        criteriaVersion: { type: string }
        teacherMessage: { type: [string, 'null'] }
        submittedAt: { type: string, format: date-time }
        checkedAt: { type: [string, 'null'], format: date-time }
        checks:
          type: array
          minItems: 0
          maxItems: 4
          items: { $ref: '#/components/schemas/TeacherPhotoCheck' }
        finalPhoto:
          oneOf:
            - type: object
              additionalProperties: false
              required: [fileId, contentUrl, filterPreset, filterStrength, savedAt]
              properties:
                fileId: { type: string, format: uuid }
                contentUrl: { type: string, format: uri }
                filterPreset:
                  type: string
                  enum: [cream_bright_04_adaptive_v1, cream_bright_04_adaptive_v2, cream_bright_04_adaptive_v3]
                filterStrength: { type: number, enum: [0.6, 0.75, 1] }
                savedAt: { type: string, format: date-time }
            - { type: 'null' }
...
    CreateSupportTicket:
      type: object
      additionalProperties: false
      required: [secondaryCategory, problemLocation, description]
      properties:
        secondaryCategory:
          type: string
          enum:
            [TASK_RULES, LESSON_INFO, SCORE_OR_REVIEW, PRODUCT_FUNCTION, ACCOUNT_LOGIN, MEDIA_UPLOAD_CAMERA, OTHER]
        problemLocation:
          type: string
          enum: [MY_TIDE, TASK, LESSON, MESSAGES, ACCOUNT, HELP, OTHER]
        description: { type: string, minLength: 1, maxLength: 5000 }
        context:
          type: string
          description: multipart 中序列化的 JSON 对象；服务端只保留白名单上下文
        images:
          type: array
          maxItems: 3
          items: { type: string, contentEncoding: binary }
    ReplySupportTicket:
      type: object
      additionalProperties: false
      required: [description, rowVersion]
      properties:
        description: { type: string, minLength: 1, maxLength: 5000 }
        rowVersion: { type: integer, minimum: 1 }
        images:
          type: array
          maxItems: 3
          items: { type: string, contentEncoding: binary }
...
    SupportTicket:
      type: object
      additionalProperties: false
      required:
        [ticketId, primaryCategory, secondaryCategory, problemLocation,
         problemContext, messages, status, unread, lastOperatorReplyAt,
         teacherReplyDeadlineAt, closeReason, closedAt, rowVersion, createdAt,
         updatedAt]
      properties:
        ticketId: { type: string, format: uuid }
        primaryCategory: { type: string, enum: [OPERATIONS, PRODUCT, OTHER] }
        secondaryCategory:
          type: string
          enum:
            [TASK_RULES, LESSON_INFO, SCORE_OR_REVIEW, PRODUCT_FUNCTION, ACCOUNT_LOGIN, MEDIA_UPLOAD_CAMERA, OTHER]
        problemLocation:
          type: string
          enum: [MY_TIDE, TASK, LESSON, MESSAGES, ACCOUNT, HELP, OTHER]
        problemContext: { type: object, additionalProperties: true }
        messages:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SupportTicketMessage' }
        status:
          type: string
          enum: [WAITING_OPERATOR, WAITING_TEACHER, CLOSED]
        unread: { type: boolean }
        lastOperatorReplyAt: { type: [string, 'null'], format: date-time }
        teacherReplyDeadlineAt: { type: [string, 'null'], format: date-time }
        closeReason:
          type: [string, 'null']
          enum: [RESOLVED, NO_RESPONSE_TIMEOUT, null]
        closedAt: { type: [string, 'null'], format: date-time }
        rowVersion: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SupportTicketList:
      type: object
      additionalProperties: false
      required: [items, unreadCount]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SupportTicket' }
        unreadCount: { type: integer, minimum: 0 }
...
    ApiError:
      type: object
      required: [code, message, requestId, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        retryable: { type: boolean }
        details: { type: object, additionalProperties: true }